requirements.txt

numpy 
torch 
torchvision
scipy
matplotlib
pathlib
pillow
jason
os-sys
scikit-image
trimesh
